Helsesjekk må bruke sts då jwt token er null for helsesjekker
