+++ src/main/kotlin/no/nav/familie/ef/søknad/config/ApplicationConfig.kt
@@ -91,7 +91,6 @@ internal class ApplicationConfig {
                 .build()
     }
 
-    // TODO skal bruke STS for å hente barn
     @Bean("stsRestKlientMedApiKey")
     fun stsRestTemplateMedApiKey(consumerIdClientInterceptor: ConsumerIdClientInterceptor,
                                  stsBearerTokenClientInterceptor: StsBearerTokenClientInterceptor,
